// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const FireIncidentStatus = {
  REPORTED: 'REPORTED',
  DISPATCHED: 'DISPATCHED',
  ON_GOING: 'ON_GOING',
  UNDER_CONTROL: 'UNDER_CONTROL',
  FIRE_OUT: 'FIRE_OUT'
} as const

export type FireIncidentStatus = (typeof FireIncidentStatus)[keyof typeof FireIncidentStatus]


export const AlarmStatus = {
  NO_ALARM: 'NO_ALARM',
  ALARM_RECEIVED: 'ALARM_RECEIVED',
  FALSE_ALARM: 'FALSE_ALARM',
  UNKNOWN: 'UNKNOWN'
} as const

export type AlarmStatus = (typeof AlarmStatus)[keyof typeof AlarmStatus]


export const CauseOfFire = {
  UNDER_INVESTIGATION: 'UNDER_INVESTIGATION',
  ELECTRICAL: 'ELECTRICAL',
  ARSON: 'ARSON',
  ACCIDENTAL: 'ACCIDENTAL',
  NATURAL: 'NATURAL',
  UNKNOWN: 'UNKNOWN',
  OTHER: 'OTHER'
} as const

export type CauseOfFire = (typeof CauseOfFire)[keyof typeof CauseOfFire]


export const FireCaseClassification = {
  ACCIDENTAL: 'ACCIDENTAL',
  NATURAL: 'NATURAL',
  ARSON: 'ARSON',
  UNDETERMINED: 'UNDETERMINED'
} as const

export type FireCaseClassification = (typeof FireCaseClassification)[keyof typeof FireCaseClassification]


export const FireCaseStatus = {
  FIRST_ALARM: 'FIRST_ALARM',
  SECOND_ALARM: 'SECOND_ALARM',
  THIRD_ALARM: 'THIRD_ALARM',
  FOURTH_ALARM: 'FOURTH_ALARM',
  FIFTH_ALARM: 'FIFTH_ALARM',
  TASK_FORCE_ALPHA: 'TASK_FORCE_ALPHA',
  TASK_FORCE_BRAVO: 'TASK_FORCE_BRAVO',
  TASK_FORCE_CHARLIE: 'TASK_FORCE_CHARLIE',
  TASK_FORCE_DELTA: 'TASK_FORCE_DELTA',
  GENERAL_ALARM: 'GENERAL_ALARM'
} as const

export type FireCaseStatus = (typeof FireCaseStatus)[keyof typeof FireCaseStatus]


export const ReportStatus = {
  DRAFT: 'DRAFT',
  SUBMITTED: 'SUBMITTED',
  UNDER_REVIEW: 'UNDER_REVIEW',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus]


export const CATEGORY_TYPE = {
  STRUCTURAL: 'STRUCTURAL',
  TRANSPORTATION: 'TRANSPORTATION',
  WILDLAND: 'WILDLAND'
} as const

export type CATEGORY_TYPE = (typeof CATEGORY_TYPE)[keyof typeof CATEGORY_TYPE]


export const OfficeType = {
  NATIONAL_HEADQUARTERS: 'NATIONAL_HEADQUARTERS',
  REGIONAL_OFFICE: 'REGIONAL_OFFICE',
  DISTRICT_FIRE_STATION: 'DISTRICT_FIRE_STATION',
  PROVINCIAL_FIRE_STATION: 'PROVINCIAL_FIRE_STATION',
  CITY_FIRE_STATION: 'CITY_FIRE_STATION',
  MUNICIPAL_FIRE_STATION: 'MUNICIPAL_FIRE_STATION'
} as const

export type OfficeType = (typeof OfficeType)[keyof typeof OfficeType]


export const ROLE = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  NHQ_ADMIN: 'NHQ_ADMIN',
  NHQ_CHIEF_INVESTIGATOR: 'NHQ_CHIEF_INVESTIGATOR',
  NHQ_INVESTIGATOR: 'NHQ_INVESTIGATOR',
  NHQ_VIEWER: 'NHQ_VIEWER',
  REGIONAL_ADMIN: 'REGIONAL_ADMIN',
  REGIONAL_CHIEF_INVESTIGATOR: 'REGIONAL_CHIEF_INVESTIGATOR',
  REGIONAL_INVESTIGATOR: 'REGIONAL_INVESTIGATOR',
  REGIONAL_VIEWER: 'REGIONAL_VIEWER',
  DISTRICT_CHIEF_INVESTIGATOR: 'DISTRICT_CHIEF_INVESTIGATOR',
  DISTRICT_INVESTIGATOR: 'DISTRICT_INVESTIGATOR',
  DISTRICT_VIEWER: 'DISTRICT_VIEWER',
  STATION_INVESTIGATOR: 'STATION_INVESTIGATOR',
  STATION_COMMEL: 'STATION_COMMEL'
} as const

export type ROLE = (typeof ROLE)[keyof typeof ROLE]


export const RANK = {
  FDIR: 'FDIR',
  FCSUPT: 'FCSUPT',
  FSSUPT: 'FSSUPT',
  FSUPT: 'FSUPT',
  FCINSP: 'FCINSP',
  FSINSP: 'FSINSP',
  FINSP: 'FINSP',
  SFO4: 'SFO4',
  SFO3: 'SFO3',
  SFO2: 'SFO2',
  SFO1: 'SFO1',
  FO3: 'FO3',
  FO2: 'FO2',
  FO1: 'FO1'
} as const

export type RANK = (typeof RANK)[keyof typeof RANK]